+++ ApplicationCode/ProjectDataModel/RimSimWellInView.h
@@ -71,7 +71,7 @@ public:
                                                                     std::vector< std::vector <cvf::Vec3d> >& pipeBranchesCLCoords,
                                                                     std::vector< std::vector <RigWellResultPoint> >& pipeBranchesCellIds) const;
 
-    void                                wellHeadTopBottomPosition(size_t frameIndex, cvf::Vec3d* top,  cvf::Vec3d* bottom);
+    void                                wellHeadTopBottomPosition(int frameIndex, cvf::Vec3d* top,  cvf::Vec3d* bottom);
     double                              pipeRadius();
     caf::PdmField<bool>                 showWell;
